daemon/nft: state the term-lowering disposition contract
`nftRulesFromTerm` was 423 lines lowering one Junos filter term to nft.
Interleaved through it were FOUR distinct dispositions for an input that
cannot be rendered faithfully, each chosen per match kind for a stated
reason, and inferable only by reading the whole function:

  A. EMIT NO RULE — a constrained address direction resolving to an empty
     positive set for this family. The term cannot match, so a rule would
     be dead.
  B. EMIT THE TOKEN VERBATIM — an unresolvable protocol / ICMP type /
     code. Invalid nft is the POINT: `nft -f -` loads atomically, so the
     whole ruleset is REJECTED and the prior generation stays installed
     (#6806, matching #6405 ports/DSCP and #6512 addresses).
  C. FAIL THE TERM CLOSED WITH A SCOPED DROP — an unrepresentable
     tcp-flags expression (#5512).
  D. MAKE THE TERM MATCH NOTHING — an unrepresentable
     flexible-match-range (#6804), mirroring userspace's
     FlexMatchStart::Unsupported. Deliberately NOT C, because a
     flex-match constraint is not necessarily protocol-scoped.

None of them is "drop the predicate and carry on" — that is the fail-open
#6405/#6512 removed from this very function, and #6806 exists because one
disposition had shipped wrong: its own comment records that the code
"previously said the token was dropped with a warning" when it does not.

A contract that is invisible is a contract that gets one member wrong.
So the three lowering steps that carry a disposition are now named for
it — nftTermAddrPredicates (A), nftTermProtocolPredicates (B),
nftTermICMPPredicates (B) — and the set is stated once, above them, with
C and D pointed at their still-inline sites.

  nftRulesFromTerm      423 -> 309 lines
  daemon_nft.go        2661 -> 2236
  daemon_nft_term_lower.go     507 (new)

The file stays ABOVE the 2000 tier and that is the correct outcome: the
tier is a trigger for asking whether a seam exists, not a number to get
under. A split sized to clear a threshold is the churn #6870 was closed
for. What was worth doing here is the god-function and its contract; the
file boundary just follows the seam.

Not pure code motion — the address block's two early `return nil` paths
became an `emitNoRule` return, so control flow moved. RUN-count equality
(1492 = 1492, measured with `git stash -u` so the base actually builds)
therefore proves only that nothing was decommissioned, NOT that the
contract survived. Mutation cells per disposition follow in the next
commit and are what carry that claim.

Boundary check: the moved region stops at nftRulesFromTerm's closing
brace, so `joinNftFields`'s godoc stayed with `joinNftFields`. My first
attempt trimmed trailing blanks and asserted a `}` — it hit that godoc
instead, which is the doc-stealing hazard catching itself.

Validation: go build ./... , go vet ./pkg/daemon/ clean, gofmt clean on
both files, go test ./pkg/daemon/ -count=1 ok.
